requirements.txt
torch>=2.1
torchaudio>=2.1
tqdm
pyyaml
numpy
soundfile
